refactor: 테스트 코드를 알맞은 패키지로 이동
