+++ src/test/java/study/lotto/domain/LottoRankTest.java
@@ -1,4 +1,4 @@
-package study.lotto;
+package study.lotto.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
